Backend/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
nanoid
pillow
fastapi[all]
azure-storage-blob[aio] 
azure-identity
azure-cosmosdb-table